Add support for setting up separate relations with custom query procs
a query proc is called to evaluate a query string before sending it do
the dataset, it can be set globally for all relations and overidden in
each relation class using `query_proc` macro.

this also makes it possible to define a directory structure where
queries are grouped per relation, ie:

  my_queries
    |- users
      |- by_name.sql
      |- by_id.sql
    |- tasks
      |- by_title.sql
      |- with_tags.sql

And it will be mapped to "Users" and "Tasks" relation automatically,
those names are derived from `dataset` setting, which means it can be
overridden.